django_cloudsql/requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
cachetools==5.3.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cloud-sql-python-connector==1.5.0
cryptography==41.0.7
Django==4.1.13
django-redis==5.4.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djongo==1.3.6
dnspython==2.4.2
eventlet==0.33.3
frozenlist==1.4.1
gevent==23.9.0.post1
google-api-core==2.15.0
google-auth==2.25.2
google-cloud-core==2.4.1
google-cloud-firestore==2.14.0
googleapis-common-protos==1.62.0
greenlet==3.0.3
grpcio==1.60.0
grpcio-status==1.60.0
gunicorn==21.2.0
idna==3.4
multidict==6.0.4
mysqlclient==2.2.0
packaging==23.1
proto-plus==1.23.0
protobuf==4.25.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pymongo==4.6.1
PyMySQL==1.1.0
pytz==2023.3
redis==5.0.7
requests==2.31.0
rsa==4.9
six==1.16.0
SQLAlchemy==2.0.23
sqlparse==0.2.4
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
waitress==2.1.2
yarl==1.9.4
zope.event==5.0
zope.interface==6.0